Fixed crashes, started work on handling Stop button, fixed a few code issues found by Coverity, fixed consistent handling of lambda paramters

+++ src/filters/deconvolution/deconvolution.h
@@ -38,7 +38,7 @@ EXTERNC typedef struct estk_data {
 	int ntries;
 	int nouter;
 	float compensationfactor;
-	float medianfilter;
+	int medianfilter;
 	float finaldeconvolutionweight;
 	float intermediatedeconvolutionweight;
 	int blindtype;
